refactor(token-provisioning): split shared token dialog ownership
