+++ src/features/KeyManagement/components/Header.tsx
@@ -7,13 +7,14 @@ import { PageHeader } from "~/components/PageHeader"
 import Tooltip from "~/components/Tooltip"
 import { Button, IconButton } from "~/components/ui"
 import { ProductAnalyticsScope } from "~/contexts/ProductAnalyticsScopeContext"
-import { KEY_MANAGEMENT_TEST_IDS } from "~/features/KeyManagement/testIds"
 import {
   PRODUCT_ANALYTICS_ACTION_IDS,
   PRODUCT_ANALYTICS_ENTRYPOINTS,
   PRODUCT_ANALYTICS_FEATURE_IDS,
   PRODUCT_ANALYTICS_SURFACE_IDS,
 } from "~/services/productAnalytics/events"
+
+import { KEY_MANAGEMENT_TEST_IDS } from "../testIds"
 
 interface HeaderProps {
   selectedAccount: string
