+++ src/features/KeyManagement/components/TokenList.tsx
@@ -15,7 +15,6 @@ import { CCSwitchExportDialog } from "~/components/CCSwitchExportDialog"
 import { ManagedSiteIcon } from "~/components/icons/ManagedSiteIcon"
 import { Badge, Button, Card, Checkbox, EmptyState } from "~/components/ui"
 import { useUserPreferencesContext } from "~/contexts/UserPreferencesContext"
-import { KEY_MANAGEMENT_TEST_IDS } from "~/features/KeyManagement/testIds"
 import { saveApiTokensToApiCredentialProfiles } from "~/features/TokenProvisioning/utils/apiCredentialProfileSaveAction"
 import { cn } from "~/lib/utils"
 import type { ManagedSiteTokenChannelStatus } from "~/services/managedSites/tokenChannelStatus"
@@ -38,6 +37,7 @@ import { createTab } from "~/utils/browser/browserApi"
 import { createLogger } from "~/utils/core/logger"
 
 import { KEY_MANAGEMENT_ALL_ACCOUNTS_VALUE } from "../constants"
+import { KEY_MANAGEMENT_TEST_IDS } from "../testIds"
 import { buildTokenIdentityKey } from "../utils"
 import { BatchCliProxyExportDialog } from "./BatchCliProxyExportDialog"
 import { ManagedSiteTokenBatchExportDialog } from "./ManagedSiteTokenBatchExportDialog"
