+++ src/features/ModelList/components/ModelKeyDialog/hooks/useModelKeyDialog.ts
@@ -2,8 +2,8 @@ import { useCallback, useEffect, useMemo, useState } from "react"
 import toast from "react-hot-toast"
 import { useTranslation } from "react-i18next"
 
-import { shouldShowOneTimeKeyDialogForCreatedToken } from "~/features/KeyManagement/utils"
 import { buildGroupDefaultTokenRequest } from "~/services/accounts/accountKeyAutoProvisioning/ensureDefaultToken"
+import { shouldShowOneTimeKeyDialogForCreatedToken } from "~/services/accounts/createdTokenSecretHandling"
 import {
   canManageDisplayAccountTokens,
   createDisplayAccountApiContext,
