+++ src/features/TokenProvisioning/testIds.ts
@@ -1,0 +1,7 @@
+export const TOKEN_PROVISIONING_TEST_IDS = {
+  addTokenDialog: "key-management-add-token-dialog",
+  addTokenSubmitButton: "key-management-add-token-submit-button",
+  oneTimeKeyCloseButton: "key-management-one-time-key-close-button",
+  oneTimeKeySaveButton: "key-management-one-time-key-save-button",
+  openApiProfilesToastButton: "key-management-open-api-profiles-toast-button",
+} as const
