+++ tests/entrypoints/options/pages/ModelList/ModelKeyDialog.test.tsx
@@ -2,8 +2,8 @@ import userEvent from "@testing-library/user-event"
 import { beforeEach, describe, expect, it, vi } from "vitest"
 
 import { SITE_TYPES } from "~/constants/siteType"
-import { KEY_MANAGEMENT_TEST_IDS } from "~/features/KeyManagement/testIds"
 import ModelKeyDialog from "~/features/ModelList/components/ModelKeyDialog"
+import { TOKEN_PROVISIONING_TEST_IDS } from "~/features/TokenProvisioning/testIds"
 import {
   PRODUCT_ANALYTICS_ACTION_IDS,
   PRODUCT_ANALYTICS_ENTRYPOINTS,
@@ -466,7 +466,9 @@ describe("ModelKeyDialog", () => {
       }),
     )
     await user.click(
-      await screen.findByTestId(KEY_MANAGEMENT_TEST_IDS.oneTimeKeySaveButton),
+      await screen.findByTestId(
+        TOKEN_PROVISIONING_TEST_IDS.oneTimeKeySaveButton,
+      ),
     )
 
     await waitFor(() => {
