+++ tests/features/KeyManagement/utils.test.ts
@@ -1,13 +1,10 @@
 import { beforeEach, describe, expect, it, vi } from "vitest"
 
-import { SITE_TYPES } from "~/constants/siteType"
 import { UI_CONSTANTS } from "~/constants/ui"
 import {
   buildTokenIdentityKey,
   formatKey,
   formatQuota,
-  shouldShowOneTimeKeyDialogForAccount,
-  shouldShowOneTimeKeyDialogForCreatedToken,
 } from "~/features/KeyManagement/utils"
 
 const { tMock } = vi.hoisted(() => ({
@@ -74,61 +71,6 @@ describe("KeyManagement utils", () => {
       const quota = UI_CONSTANTS.EXCHANGE_RATE.CONVERSION_FACTOR * 1.25
 
       expect(formatQuota(quota, false)).toBe("$1.25")
-    })
-  })
-
-  describe("shouldShowOneTimeKeyDialogForAccount", () => {
-    it("shows the one-time key dialog for AIHubMix create responses", () => {
-      expect(
-        shouldShowOneTimeKeyDialogForAccount({
-          siteType: SITE_TYPES.AIHUBMIX,
-        }),
-      ).toBe(true)
-    })
-
-    it("does not force a one-time key dialog for Sub2API create responses", () => {
-      expect(
-        shouldShowOneTimeKeyDialogForAccount({
-          siteType: SITE_TYPES.SUB2API,
-        }),
-      ).toBe(false)
-    })
-
-    it("does not force a one-time key dialog for New API create responses", () => {
-      expect(
-        shouldShowOneTimeKeyDialogForAccount({
-          siteType: SITE_TYPES.NEW_API,
-        }),
-      ).toBe(false)
-    })
-  })
-
-  describe("shouldShowOneTimeKeyDialogForCreatedToken", () => {
-    it("requires an AIHubMix account and a usable unmasked secret", () => {
-      expect(
-        shouldShowOneTimeKeyDialogForCreatedToken(
-          { siteType: SITE_TYPES.AIHUBMIX },
-          { key: "sk-aihubmix-full-secret" },
-        ),
-      ).toBe(true)
-    })
-
-    it("does not show the dialog for masked AIHubMix created keys", () => {
-      expect(
-        shouldShowOneTimeKeyDialogForCreatedToken(
-          { siteType: SITE_TYPES.AIHUBMIX },
-          { key: "sk-aihub********masked" },
-        ),
-      ).toBe(false)
-    })
-
-    it("does not show the dialog for full Sub2API created keys", () => {
-      expect(
-        shouldShowOneTimeKeyDialogForCreatedToken(
-          { siteType: SITE_TYPES.SUB2API },
-          { key: "sk-sub2api-full-secret" },
-        ),
-      ).toBe(false)
     })
   })
 })
