+++ tests/features/TokenProvisioning/utils/apiCredentialProfileSaveAction.test.ts
@@ -6,7 +6,7 @@ import {
   AIHUBMIX_WEB_ORIGIN,
   SITE_TYPES,
 } from "~/constants/siteType"
-import { KEY_MANAGEMENT_TEST_IDS } from "~/features/KeyManagement/testIds"
+import { TOKEN_PROVISIONING_TEST_IDS } from "~/features/TokenProvisioning/testIds"
 import {
   buildOneTimeApiKeyProfileSaveAction,
   saveApiTokensToApiCredentialProfiles,
@@ -299,7 +299,7 @@ describe("saveApiTokensToApiCredentialProfiles", () => {
     const toastMessage = toastMessageRenderer({ id: "toast-1" })
     const openButton = toastMessage.props.children[1]
     expect(openButton.props["data-testid"]).toBe(
-      KEY_MANAGEMENT_TEST_IDS.openApiProfilesToastButton,
+      TOKEN_PROVISIONING_TEST_IDS.openApiProfilesToastButton,
     )
     expect(openButton.props.children).toBe(
       "keyManagement:actions.openApiProfiles",
